Hold all build code in src folder
