+++ src/assets/animations/jl-logo-animation.json
