+++ src/assets/animations/jl-nav-logo-animation.json
