+++ src/styles/about/about.scss
