+++ src/styles/components/navbar.scss
