+++ src/styles/components/project-item.scss
