+++ src/styles/loading-page.scss
